プロジェクトのオーディオとUIセットアップの大幅な変更
- `CG2_DirectXGame.vcxproj` に `$(ProjectDir)engine/audio` を追加
- `MagAudio` 関連ファイルを `MAudioG` にリネーム
- `ImguiSetup` 関連ファイルを `engine/base` に移動
- `ImguiSetup.cpp` と `ImguiSetup.h` の内容を削除
- `MagAudio.cpp` と `MagAudio.h` の内容を削除
- `imgui.ini` のウィンドウサイズと位置を変更
- `main.cpp` にオーディオ初期化と再生コードを追加
- `MAudioG.cpp` と `MAudioG.h` に新しいオーディオ機能を追加
- `ImguiSetup.cpp` と `ImguiSetup.h` に新しいUIセットアップ機能を追加
