+++ imgui.ini
@@ -14,7 +14,7 @@ Size=301,724
 
 [Window][2D Object]
 Pos=1,261
-Size=500,385
+Size=500,275
 
 [Window][3D Object]
 Pos=1070,39
@@ -43,6 +43,6 @@ Pos=-1,670
 Size=197,53
 
 [Window][Performance Monitor]
-Pos=503,347
+Pos=8,539
 Size=416,181
